requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1